project48/requirements.txt
asgiref==3.5.2
astroid==2.12.5
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
distlib==0.3.6
Django==4.1
django-cleanup==6.0.0
django-debug-toolbar==3.6.0
django-filter==22.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
extends==0.3.0
Faker==12.0.1
filelock==3.8.0
Flask==2.2.2
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mixer==7.2.2
phonenumbers==8.12.55
Pillow==9.2.0
pipenv==2022.8.31
platformdirs==2.5.2
pylint==2.15.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
urllib3==1.26.12
virtualenv==20.16.4
virtualenv-clone==0.5.7
Werkzeug==2.2.2
wrapt==1.14.1
